Make ping() a bit more intelligent; fix sudo aliases
ping() now checks to see if the local machine has a valid IPv6
address(in the global scope) before actually calling ping6.

No longer add files which are both in /sbin and /bin to the sbin aliases
function. This was causing a stupid interaction with `sudo ping6`,
because it is a setuid program and symlinked to /sbin on certain
systems.